refactor: Extract buildQuery from DriveView
Will be easier to reuse in other components
