+++ src/drive/web/modules/views/Drive/index.jsx
@@ -1,6 +1,7 @@
 /* global __TARGET__ */
 import React, { useCallback, useState, useContext } from 'react'
-import { useQuery, Q } from 'cozy-client'
+import { connect } from 'react-redux'
+import { useQuery } from 'cozy-client'
 
 import SharingProvider from 'cozy-sharing'
 import {
@@ -14,7 +15,7 @@ import Dropzone from 'drive/web/modules/upload/Dropzone'
 import Main from 'drive/web/modules/layout/Main'
 import Topbar from 'drive/web/modules/layout/Topbar'
 import Toolbar from 'drive/web/modules/drive/Toolbar'
-import { ROOT_DIR_ID, TRASH_DIR_ID } from 'drive/constants/config'
+import { ROOT_DIR_ID } from 'drive/constants/config'
 
 import { FileListv2 } from 'drive/web/modules/filelist/FileList'
 import { ConnectedFileListBodyV2 as FileListBodyV2 } from 'drive/web/modules/filelist/FileListBody'
@@ -28,25 +29,7 @@ import { isMobileApp } from 'cozy-device-helper'
 import LoadMore from 'drive/web/modules/filelist/LoadMoreV2'
 import Breadcrumb from './Breadcrumb'
 import File from './FileWithActions'
-
-const buildQuery = ({ currentFolderId, type, sortAttribute, sortOrder }) => ({
-  definition: () =>
-    Q('io.cozy.files')
-      .where({
-        dir_id: currentFolderId,
-        _id: { $ne: TRASH_DIR_ID },
-        type
-      })
-      .indexFields(['dir_id', 'type', sortAttribute])
-      .sortBy([
-        { dir_id: sortOrder },
-        { type: sortOrder },
-        { [sortAttribute]: sortOrder }
-      ]),
-  options: {
-    as: `${type}-${currentFolderId}-${sortAttribute}-${sortOrder}`
-  }
-})
+import { buildQuery } from 'drive/web/modules/queries'
 
 const DriveView = ({ params, router, children }) => {
   const { isBigThumbnail, toggleThumbnailSize } = useContext(
